ahk: Adds Logitech relaunch on mouse back
This is specific to my logitech mouse when using a USB switch. When switching the devices back, the Logitech software fails to notice the device has been "plugged back in" and the button remappings do not go into effect until the software is restarted. To remedy this, I added this remapping of the mouse back button to relaunch the software. Since I have this button remapped in the software, the relaunch is only triggered when the software has malfunctioned.